[#2149] - Agrupa los fixtures de multimedia en su carpeta
Los medios no son de ninguna de las dos entidades: la story y la obra
literaria del mismo slug consumen el mismo objeto, y duplicarlo rompería
la invariante de que ambas caras declaren exactamente los mismos. Por eso
van a una carpeta propia en vez de a la de alguna de las dos.

Extraerlos elimina de paso un import cruzado: el crudo de la obra tiraba
del crudo de la story para no duplicar el array, y ahora las dos tiran del
mismo lugar.
