+++ src/mocks/onoff/geometria.mock.ts
@@ -1,4 +1,3 @@
-import type { Media } from '@models/media.model';
 import type { Story } from '@models/story.model';
 import { createLiteraryWork, type LiteraryWork } from '@models/literary-work.model';
 import { createAttributedText } from '@models/attributed-text.model';
@@ -16,47 +15,7 @@ import { authorMock } from '../author.mock';
 import { toDomainTags } from '../onoff-tags.mock';
 import { geometriaRawStory } from './geometria.raw.mock';
 import { geometriaRawLiteraryWork } from './geometria.literary-work.raw.mock';
-import {
-	geometriaAudioDescription,
-	geometriaSpaceDescription,
-	geometriaSpotifyDescription,
-	geometriaYoutubeDescription,
-} from './geometria.media';
-
-// Espeja los mediaSources del fixture raw homónimo, sin el pdfLink que el ACL descarta. Lo comparten
-// la Story y la obra literaria del mismo slug, que en el crudo declaran exactamente los mismos medios.
-export const geometriaMediaMock: Media[] = [
-	{
-		title: 'Lectura de "Geometría" por su autor',
-		type: 'audioRecording',
-		description: markdownToSanitizedHtml(createMarkdown(geometriaAudioDescription)),
-		data: { url: 'https://cdn.example.org/onoff/geometria.ogg' },
-	},
-	{
-		title: 'Conversación sobre el insomnio y la medida del tiempo',
-		type: 'spaceRecording',
-		description: markdownToSanitizedHtml(createMarkdown(geometriaSpaceDescription)),
-		data: {
-			url: 'https://cdn.example.org/onoff/geometria-space.ogg',
-			duration: '48:12',
-			hostName: 'Biblioteca del Méridien',
-			hostAvatar: 'https://cdn.example.org/onoff/biblioteca-meridien-avatar.png',
-			date: '1974-06-12',
-		},
-	},
-	{
-		title: 'Episodio dedicado a "Geometría"',
-		type: 'spotifyPodcastEpisode',
-		description: markdownToSanitizedHtml(createMarkdown(geometriaSpotifyDescription)),
-		data: { url: 'https://open.spotify.com/embed/episode/geometria' },
-	},
-	{
-		title: 'Video ensayo sobre las coordenadas del desvelo',
-		type: 'youTubeVideo',
-		description: markdownToSanitizedHtml(createMarkdown(geometriaYoutubeDescription)),
-		data: { videoId: 'geometriaVideoId' },
-	},
-];
+import { geometriaMediaMock } from './media/geometria.media.mock';
 
 export const geometriaStoryMock: Story = {
 	_id: 'onoff-story-geometria',
